==> notebooks/evaluation/4_which_is_the_best_model.ipynb <==
from pathlib import Path

import pandas as pd

processed_path = Path("../../processed")
models_leaderboard_cache_file = processed_path / "models_leaderboard_cache.csv"

models_leaderboard = pd.read_csv(models_leaderboard_cache_file)

# Which model type is on average the best choice

We use Random Search with a budget of 50 trials to optimize the hyperparameters of the models. For this, we use autogluon`s default search grid of the following models
- FastAI Neural Network
- Random Forest
- Extremely Randomized Trees

models_leaderboard.query("cleaner_type=='ConformalAutoGluon'").pipe(
    lambda df: df.value_counts(["model_type"]) / len(df) * 100,
)

models_leaderboard.query("cleaner_type=='AutoGluon'").pipe(lambda df: df.value_counts(["model_type"]) / len(df) * 100)

models_leaderboard.groupby(["model_type", "cleaner_type"])["fit_time"].describe()
